More caching, add custom-vars.json

+++ script/cache.sh
@@ -11,7 +11,7 @@ if mountpoint "/tmp/vagrant-cache"; then
     rm -rf /var/cache/yum
     ln -sf /tmp/vagrant-cache/yum /var/cache/yum
     sed -i 's/keepcache=0/keepcache=1/g' /etc/yum.conf
-    cat >>/etc/yum.conf <<EOF
+    grep -qF 'metadata_expire=' /etc/yum.conf || cat >>/etc/yum.conf <<EOF
 
 metadata_expire=90m
 mirrorlist_expire=90m
@@ -25,9 +25,10 @@ EOF
     grep -qF 'cache=' /home/${USERNAME}/.npmrc || cat >>/home/${USERNAME}/.npmrc <<EOF
 cache=/tmp/vagrant-cache/npm
 EOF
+
 fi
 
-if [ -n "$yum_proxy" ] && curl --fail -s -o /dev/null "${yum_proxy}acng-report.html"; then
+if [ -n "$yum_proxy" ] && ! grep -qF 'proxy=' /etc/yum.conf && curl --fail -s -o /dev/null "${yum_proxy}acng-report.html"; then
     echo "==> Configuring temporary Yum proxy at $yum_proxy"
     cat >>/etc/yum.conf <<EOF
 proxy=$yum_proxy
